proj2/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
bcrypt == 4.0.1
passlib[bcrypt] == 1.7.4
python-jose[cryptography]
psycopg[binary]
mistralai>=1.0.0
PyPDF2
python-multipart
aiofiles
python-dotenv
